# oa-console/openapi/oa-app.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:18400",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/api/v1/org/employees/{employeeId}": {
      "put": {
        "tags": [
          "employee-controller"
        ],
        "operationId": "update_1",
        "parameters": [
          {
            "name": "employeeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEmployee"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/org/employees/{employeeId}/reporting-line": {
      "put": {
        "tags": [
          "employee-controller"
        ],
        "operationId": "setReportingLine",
        "parameters": [
          {
            "name": "employeeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetReportingLine"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/org/employees": {
      "post": {
        "tags": [
          "employee-controller"
        ],
        "operationId": "create_1",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEmployee"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultLong"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/org/employees/{employeeId}/transfer": {
      "post": {
        "tags": [
          "employee-controller"
        ],
        "operationId": "transfer",
        "parameters": [
          {
            "name": "employeeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransferEmployee"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/org/employees/{employeeId}/leave": {
      "post": {
        "tags": [
          "employee-controller"
        ],
        "operationId": "leave",
        "parameters": [
          {
            "name": "employeeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "leaveDate",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/org/employees/{employeeId}/assignments": {
      "post": {
        "tags": [
          "employee-controller"
        ],
        "operationId": "addAssignment",
        "parameters": [
          {
            "name": "employeeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddAssignment"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultLong"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/report/headcount": {
      "get": {
        "tags": [
          "report-controller"
        ],
        "operationId": "headcount",
        "parameters": [
          {
            "name": "orgPath",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultListMapStringObject"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/report/audit": {
      "get": {
        "tags": [
          "report-controller"
        ],
        "operationId": "auditQuery",
        "parameters": [
          {
            "name": "actorId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "action",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "deniedOnly",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultListMapStringObject"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/report/attendance": {
      "get": {
        "tags": [
          "report-controller"
        ],
        "operationId": "attendance",
        "parameters": [
          {
            "name": "days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 30
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultListMapStringObject"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/org/employees/by-user/{userId}": {
      "get": {
        "tags": [
          "employee-controller"
        ],
        "operationId": "byUser",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultEmployeeView"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/org/employees/by-user/{userId}/manager-chain": {
      "get": {
        "tags": [
          "employee-controller"
        ],
        "operationId": "managerChain",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "maxLevel",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 5
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultListString"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/org/employees/by-user/{userId}/assignments": {
      "get": {
        "tags": [
          "employee-controller"
        ],
        "operationId": "assignments",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultListAssignmentView"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/org/employees/by-user/{userId}/as-of": {
      "get": {
        "tags": [
          "employee-controller"
        ],
        "operationId": "asOf",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultOrgSnapshotView"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/org/employees/assignments/{assignmentId}": {
      "delete": {
        "tags": [
          "employee-controller"
        ],
        "operationId": "closeAssignment",
        "parameters": [
          {
            "name": "assignmentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "validTo",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ResultVoid": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "object"
          },
          "traceId": {
            "type": "string"
          }
        }
      },
...
      "UpdateEmployee": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "enName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "mobile": {
            "type": "string"
          },
          "avatar": {
            "type": "string"
          },
          "gender": {
            "type": "string"
          },
          "birthday": {
            "type": "string",
            "format": "date"
          }
        }
      },
      "SetReportingLine": {
        "required": [
          "managerEmployeeId",
          "type"
        ],
        "type": "object",
        "properties": {
          "managerEmployeeId": {
            "type": "integer",
            "format": "int64"
          },
          "type": {
            "type": "string"
          },
          "validFrom": {
            "type": "string",
            "format": "date"
          }
        }
      },
...
      "ResultLong": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "integer",
            "format": "int64"
          },
          "traceId": {
            "type": "string"
          }
        }
      },
      "ResultListString": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "traceId": {
            "type": "string"
          }
        }
      },
      "CreateEmployee": {
        "required": [
          "empNo",
          "name",
          "primaryOrgId",
          "userId"
        ],
        "type": "object",
        "properties": {
          "userId": {
            "type": "string"
          },
          "empNo": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "enName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "mobile": {
            "type": "string"
          },
          "idCard": {
            "type": "string"
          },
          "gender": {
            "type": "string"
          },
          "birthday": {
            "type": "string",
            "format": "date"
          },
          "hireDate": {
            "type": "string",
            "format": "date"
          },
          "employmentType": {
            "type": "string"
          },
          "primaryOrgId": {
            "type": "integer",
            "format": "int64"
          },
          "primaryPositionId": {
            "type": "integer",
            "format": "int64"
          },
          "managerEmployeeId": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "TransferEmployee": {
        "required": [
          "effectiveDate",
          "targetOrgId"
        ],
        "type": "object",
        "properties": {
          "targetOrgId": {
            "type": "integer",
            "format": "int64"
          },
          "targetPositionId": {
            "type": "integer",
            "format": "int64"
          },
          "effectiveDate": {
            "type": "string",
            "format": "date"
          },
          "asLeader": {
            "type": "boolean"
          },
          "reason": {
            "type": "string"
          }
        }
      },
      "AddAssignment": {
        "required": [
          "assignmentType",
          "orgUnitId"
        ],
        "type": "object",
        "properties": {
          "orgUnitId": {
            "type": "integer",
            "format": "int64"
          },
          "positionId": {
            "type": "integer",
            "format": "int64"
          },
          "assignmentType": {
            "type": "string"
          },
          "asLeader": {
            "type": "boolean"
          },
          "validFrom": {
            "type": "string",
            "format": "date"
          }
        }
      },
...
      "ResultListMapStringObject": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "additionalProperties": {
                "type": "object"
              }
            }
          },
          "traceId": {
            "type": "string"
          }
        }
      },
...
      "EmployeeView": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "userId": {
            "type": "string"
          },
          "empNo": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "enName": {
            "type": "string"
          },
          "avatar": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "employmentType": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "hireDate": {
            "type": "string",
            "format": "date"
          },
          "primaryOrgId": {
            "type": "integer",
            "format": "int64"
          },
          "primaryOrgName": {
            "type": "string"
          },
          "primaryOrgPath": {
            "type": "string"
          },
          "primaryPositionId": {
            "type": "integer",
            "format": "int64"
          },
          "primaryPositionName": {
            "type": "string"
          }
        }
      },
      "ResultEmployeeView": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/EmployeeView"
          },
          "traceId": {
            "type": "string"
          }
        }
      },
      "AssignmentView": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "employeeId": {
            "type": "integer",
            "format": "int64"
          },
          "orgUnitId": {
            "type": "integer",
            "format": "int64"
          },
          "orgName": {
            "type": "string"
          },
          "orgPath": {
            "type": "string"
          },
          "positionId": {
            "type": "integer",
            "format": "int64"
          },
          "positionName": {
            "type": "string"
          },
          "assignmentType": {
            "type": "string"
          },
          "leader": {
            "type": "boolean"
          },
          "validFrom": {
            "type": "string",
            "format": "date"
          },
          "validTo": {
            "type": "string",
            "format": "date"
          }
        }
      },
      "ResultListAssignmentView": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AssignmentView"
            }
          },
          "traceId": {
            "type": "string"
          }
        }
      },
      "OrgSnapshotView": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string"
          },
          "asOf": {
            "type": "string",
            "format": "date"
          },
          "primaryOrgId": {
            "type": "integer",
            "format": "int64"
          },
          "primaryOrgPath": {
            "type": "string"
          },
          "primaryOrgName": {
            "type": "string"
          },
          "allOrgIds": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            }
          },
          "pathPrefixes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "managerUserId": {
            "type": "string"
          }
        }
      },
      "ResultOrgSnapshotView": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/OrgSnapshotView"
          },
          "traceId": {
            "type": "string"
          }
        }
      },
...
    }
  }
}